requirements.txt
PySide6>=6.5.0
python-can>=4.0.0
udsoncan>=1.20.0
can-isotp>=2.0.0
pyserial>=3.5
